Task 3: Xử lý đồng bộ dữ liệu, thêm test và logic đấu giá

+++ server/src/main/java/com/auction/server/dto/BidResponse.java
@@ -1,21 +1,28 @@
 package com.auction.server.dto;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 public class BidResponse implements Serializable {
     private static final long serialVersionUID = 1L;
 
     private boolean success;
     private String message;
-    private double currentPrice;
+    private BigDecimal currentPrice;
 
-    public BidResponse(boolean success, String message, double currentPrice) {
+    public BidResponse(boolean success, String message, BigDecimal currentPrice) {
         this.success = success;
         this.message = message;
         this.currentPrice = currentPrice;
     }
 
-    public boolean isSuccess() { return success; }
-    public String getMessage() { return message; }
-    public double getCurrentPrice() { return currentPrice; }
+    public boolean isSuccess() {
+        return success;
+    }
+    public String getMessage() {
+        return message;
+    }
+    public BigDecimal getCurrentPrice() {
+        return currentPrice;
+    }
 }
